backend/requirements.txt
cryptography == 40.0.0
fastapi  == 0.95.0
mysql-connector == 2.2.9
mysqlclient == 2.1.1
passlib == 1.7.4
pydantic == 1.10.7
PyMySQL == 1.0.2
python-jose == 3.3.0
python-multipart == 0.0.6
setuptools == 63.2.0
SQLAlchemy == 2.0.7
starlette == 0.26.1
typing_extensions == 4.5.0
uvicorn == 0.21.1
python-decouple
requests